backend/requirements.txt
# FastAPI 核心
fastapi==0.109.0
uvicorn[standard]==0.27.0

# 异步 HTTP 客户端
httpx==0.26.0

# 文件上传支持
python-multipart==0.0.6

# 异步文件操作
aiofiles==23.2.1

# 类型验证
pydantic==2.5.3

# 环境变量
python-dotenv==1.0.0

# 日志
loguru>=0.7.0

# 图片处理
Pillow>=10.0.0
